+++ NEWS.md
@@ -1,7 +1,8 @@
 # xxdi 1.1.0
 
-* Added function for calculating novel xc-index.
-* Added functions for calculating h-index and g-index.
+* Added function 'xc_index' for calculating xc-index.
+* Added a generic function 'xdc_index' which can be used to calculate x-index, xd-index, and xc-index.
+* Added functions 'h_index' and 'g_index' for calculating h-index and g-index respectively.
 * Optimised functions for better performance and faster run times.
 
 --------------------------------------------------------------------------------